tools/coverage-html: refactor pass — assets, modules, helpers, tests
Mechanical refactor; no user-visible behavior change. Verified by
running the full 96-test suite and rendering the merged real-data
LCOV — output is byte-identical except for the timestamp and a
single `</script><script>` boundary that collapsed when the two JS
IIFEs moved into one file.

Concrete changes:

1. **CSS extracted** to `tools/coverage-html/style.css` (230 lines).
   The Python `INLINE_CSS` triple-string is gone; the renderer
   loads the file at module init and embeds it inline in each
   page. CSS linters / editors finally see real CSS.

2. **JS extracted** to `tools/coverage-html/script.js` (133 lines)
   — both IIFEs (chrome-height measurement + dir/file toggle
   handlers). Same loader pattern. The `\\u25BC` / `\\u25B6`
   escapes that were awkward in Python source are now plain
   `\u25BC` / `\u25B6` in JS. Per-file pages now also load the
   toggle IIFE (a no-op there since `.fnToggle` / `.dirToggle`
   don't exist on per-file pages); this lets us ship one JS
   file instead of two.

3. **`SourceResolver` moved** from the renderer to `lcov_io`. It
   has no UI dependency; future tools (the merger, a diff CLI)
   may need it.

4. **`_hue_for_pct(pct)`** factored out so `_gradient_color` and
   `_gradient_cell_bg` share the formula.

5. **Style constants centralized** in a "Style constants" block at
   the top of the renderer: `TIER_HI`, `TIER_MED`, `BAR_PIXEL_WIDTH`,
   `GRADIENT_HUE_PER_PCT`, the four gradient saturation/lightness
   pairs, `BRANCH_COL_WIDTH`, `CHEVRON_OPEN`/`CHEVRON_CLOSED`. Magic
   numbers no longer scattered across the file.

6. **`render_index` (273 lines) split** into focused helpers:
   `_index_dir_tree`, `_files_under`, `_dir_depth`, `_indent_style`,
   `_index_thead_html`, `_index_colgroup_html`,
   `_render_dir_header_row`, `_render_file_row`,
   `_render_file_functions_row`. The orchestrator drops to ~70
   lines.

7. **`_render_inline_functions_table` (136 lines) split** into
   `_fn_line_coverage_cells`, `_fn_function_coverage_cells`,
   `_fn_branch_coverage_cells`, `_fn_inner_colgroup`, and
   `_fn_inner_thead`. Body becomes a clean per-row composition.

8. **Tests split** between `test_lcov_io.py` (parser, FileRecord,
   `function_line_coverage`, `function_branch_coverage`,
   `_effective_fn_hit`, `apply_slangc_filter`, `SourceResolver`) and
   `test_renderer.py` (HTML rendering, gradient/tier helpers, CLI
   integration). `test_merge.py` unchanged.

Sizes after refactor:

  before:                              after:
    slang-coverage-html.py: 1654         slang-coverage-html.py: 1293
    lcov_io.py:              585         lcov_io.py:              647
    style.css:               (inline)    style.css:               230
    script.js:               (inline)    script.js:               133
    tests/test_renderer.py: 1223         tests/test_renderer.py:  795
    tests/test_lcov_io.py:   (n/a)       tests/test_lcov_io.py:   492
    tests/test_merge.py:     305         tests/test_merge.py:     305

The renderer drops 361 LOC (mostly inline CSS + JS that moved to
sibling files). Splitting the two largest renderer functions tames
the navigation cost — every file is comfortably under ~700 LOC.

96 tests still pass. Output equivalence verified against a baseline
render of the PR-10884 3-OS slangc merge.
